♻️ refactor: refactor the ModelSelect
